consolidate shouldTrackFill and shouldTrackKeyPress into resolveAction
- remove old selector code we will replace
