+++ recorder/src/element.ts
@@ -1,20 +1,50 @@
 import { getXpath } from "./serialize";
 
-type OnElementAddedToGroupFn = (element: HTMLElement, depth: number) => void;
+// --
+// from playwright to match their text engine
+const cachedText = new Map<Element | ShadowRoot, string>();
 
-type TraverseClickableElementsInput = {
-  ancestorChain?: string[];
-  depth?: number;
-  direction?: "up" | "down";
-  element: HTMLElement;
-  group: HTMLElement[];
-  maxDepth?: number;
-  onElementAddedToGroup?: OnElementAddedToGroupFn;
-};
+function shouldSkipForTextMatching(element: Element | ShadowRoot) {
+  return (
+    element.nodeName === "SCRIPT" ||
+    element.nodeName === "STYLE" ||
+    (document.head && document.head.contains(element))
+  );
+}
 
-const BUTTON_INPUT_TYPES = ["button", "image", "reset", "submit"];
-const CLICK_GROUP_ELEMENTS = ["a", "button", "label"];
-const MAX_CLICKABLE_ELEMENT_TRAVERSE_DEPTH = 10;
+export function elementText(root: Element | ShadowRoot): string {
+  let value = cachedText.get(root);
+  if (value === undefined) {
+    value = "";
+    if (!shouldSkipForTextMatching(root)) {
+      if (
+        root instanceof HTMLInputElement &&
+        (root.type === "submit" || root.type === "button")
+      ) {
+        value = root.value;
+      } else {
+        for (let child = root.firstChild; child; child = child.nextSibling) {
+          if (child.nodeType === Node.ELEMENT_NODE)
+            value += elementText(child as Element);
+          else if (child.nodeType === Node.TEXT_NODE)
+            value += child.nodeValue || "";
+
+          // skip long text
+          if (value.length > 100) {
+            value = "";
+            break;
+          }
+        }
+        if ((root as Element).shadowRoot)
+          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+          value += elementText((root as Element).shadowRoot!);
+      }
+    }
+    cachedText.set(root, value);
+  }
+  return value;
+}
+// --
 
 export const isVisible = (
   element: Element,
@@ -34,133 +64,6 @@ export const isVisible = (
   }
 
   return true;
-};
-
-export const isLikelyTopOfClickGroup = (element: HTMLElement): boolean => {
-  const tagName = element.tagName.toLowerCase();
-  return (
-    CLICK_GROUP_ELEMENTS.includes(tagName) ||
-    element.getAttribute("role") === "button" ||
-    (tagName === "input" &&
-      BUTTON_INPUT_TYPES.includes(element.getAttribute("type")))
-  );
-};
-
-/**
- * @summary Traverse the DOM in both directions, adding clickable elements to the array
- *   until we've determined the full set of elements that are in the clickable area.
- *   This is not foolproof because we can't know where exactly click handlers might
- *   be attached, but we can do a pretty good job of guessing.
- */
-const traverseClickableElements = (
-  input: TraverseClickableElementsInput
-): void => {
-  const {
-    ancestorChain = [],
-    depth = 0,
-    direction = "up",
-    element,
-    group,
-    maxDepth = MAX_CLICKABLE_ELEMENT_TRAVERSE_DEPTH,
-    onElementAddedToGroup,
-  } = input;
-
-  // Regardless of which direction we're moving, stop if we hit an invisible element
-  if (!isVisible(element, window.getComputedStyle(element))) return;
-
-  // When moving up, when we reach the topmost clickable element, we
-  // stop traversing up and begin traversing down from there.
-  if (direction === "up" && isLikelyTopOfClickGroup(element)) {
-    traverseClickableElements({
-      direction: "down",
-      element,
-      group,
-      maxDepth,
-      onElementAddedToGroup,
-    });
-    return;
-  }
-
-  // When moving down, stop if we hit the top of another click group (nested buttons)
-  if (direction === "down" && depth > 0 && isLikelyTopOfClickGroup(element))
-    return;
-
-  const newDepth = depth + 1;
-  const lowerTagName = element.tagName.toLowerCase();
-
-  if (direction === "up") {
-    // Call self for the parent element, incrementing depth
-    if (element.parentElement) {
-      traverseClickableElements({
-        ancestorChain: [lowerTagName, ...ancestorChain],
-        direction,
-        element: element.parentElement,
-        group,
-        maxDepth,
-        depth: newDepth,
-        onElementAddedToGroup,
-      });
-    }
-  } else {
-    // Respect max depth only when going down
-    if (newDepth > maxDepth) return;
-
-    // If we make it this far, this element should be part of the current group.
-    // We add elements to the group only on the way down to avoid adding any twice.
-    group.push(element);
-
-    // Let caller do additional things with each element as we add it
-    if (onElementAddedToGroup) onElementAddedToGroup(element, depth);
-
-    const newAncestorChain = [...ancestorChain, lowerTagName];
-    console.debug(
-      "qawolf: added %s to click group",
-      newAncestorChain.join(" > ")
-    );
-
-    // For now, let's skip going down into SVG descendants to keep this fast. If anyone
-    // finds a situation in which this causes problems, we can remove this.
-    if (lowerTagName !== "svg") {
-      for (const child of element.children) {
-        // Call self for each child element, incrementing depth
-        traverseClickableElements({
-          ancestorChain: newAncestorChain,
-          direction,
-          element: child as HTMLElement,
-          group,
-          maxDepth,
-          depth: newDepth,
-          onElementAddedToGroup,
-        });
-      }
-    }
-  }
-};
-
-/**
- * @summary Sometimes there is a group of elements that together make up what appears
- *   to be a single button, link, image, etc. Examples: a > span | button > div > span
- *   For these we want to take into consideration the entire "clickable group" when
- *   building a good selector. The topmost clickable (a | button | input) should be
- *   preferred in many cases, but if an inner element has a lower-penalty attribute
- *   then that should be preferred.
- *
- * @return An array of HTMLElement that make up the clickable group. If `element`
- *   itself is not clickable, the array is empty.
- */
-export const getClickableGroup = (
-  element: HTMLElement,
-  onElementAddedToGroup?: OnElementAddedToGroupFn
-): HTMLElement[] => {
-  console.debug("qawolf: get clickable ancestor for", getXpath(element));
-
-  const group = [];
-
-  // Recursive function that will mutate clickableElements array. A recursive
-  // function is better than loops to avoid blocking UI paint.
-  traverseClickableElements({ element, group, onElementAddedToGroup });
-
-  return group;
 };
 
 /**
