+++ recorder/src/qawolf.ts
@@ -1,19 +1,10 @@
-export { getAttribute, getRegexAttribute } from "./attribute";
 export { resolveAction } from "./resolveAction";
-export {
-  buildCues,
-  buildCuesForElement,
-  buildCueValueForTag,
-  getCueTypesConfig,
-} from "./cues";
 export { getCues, getTagValue } from "./getCues";
 export {
-  getClickableGroup,
   getInputElementValue,
   getTopmostEditableElement,
   isVisible,
 } from "./element";
 export { ActionRecorder } from "./ActionRecorder";
-export { buildSelector, clearSelectorCache } from "./selector";
 export { buildElementText, isMatch } from "./selectorEngine";
 export { getXpath, nodeToDoc } from "./serialize";
