+++ recorder/src/types.ts
@@ -17,6 +17,7 @@ export type Callback<S = void, T = void> = (data?: S) => T;
 export type Cue = {
   level: number; // 0 is target, 1 is parent, etc.
   penalty: number; // Cue type penalty plus PENALTY_PER_LEVEL
+  type: "attribute" | "class" | "id" | "tag" | "text";
   value: string;
 };
